requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-SQLAlchemy
